requirements.txt
asgiref==3.11.0
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
dotenv==0.9.9
gunicorn==23.0.0
packaging==25.0
pillow==12.0.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.3
tzdata==2025.2
